improved menu interface, added return buttons, pin refreshes every 5 seconds
