+++ FindMe/AccountViewController.swift
@@ -1,0 +1,39 @@
+//
+//  AccountViewController.swift
+//  FindMe
+//
+//  Created by Jordi Turner on 4/1/16.
+//  Copyright © 2016 William Tong. All rights reserved.
+//
+
+import UIKit
+
+class AccountViewController: UIViewController {
+
+    @IBOutlet weak var menuButton: UIBarButtonItem!
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        if self.revealViewController() != nil {
+            menuButton.target = self.revealViewController()
+            menuButton.action = #selector(SWRevealViewController.revealToggle(_:))
+        }
+        // Do any additional setup after loading the view.
+    }
+
+    override func didReceiveMemoryWarning() {
+        super.didReceiveMemoryWarning()
+        // Dispose of any resources that can be recreated.
+    }
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepareForSegue(segue: UIStoryboardSegue, sender: AnyObject?) {
+        // Get the new view controller using segue.destinationViewController.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
