+++ FindMe/SettingsViewController.swift
@@ -1,0 +1,31 @@
+//
+//  SettingsViewController.swift
+//  FindMe
+//
+//  Created by Jordi Turner on 4/1/16.
+//  Copyright © 2016 William Tong. All rights reserved.
+//
+
+import UIKit
+
+class SettingsViewController: UIViewController {
+
+    @IBOutlet weak var menuButton: UIBarButtonItem!
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        if self.revealViewController() != nil {
+            menuButton.target = self.revealViewController()
+            menuButton.action = #selector(SWRevealViewController.revealToggle(_:))
+        }
+        // Do any additional setup after loading the view.
+    }
+
+    override func didReceiveMemoryWarning() {
+        super.didReceiveMemoryWarning()
+        // Dispose of any resources that can be recreated.
+    }
+    
+
+    
+
+}
